# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GPU Cluster Platform",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/nodes/{node_id}/install-ssh-pubkey": {
      "post": {
        "summary": "Install Node Ssh Pubkey",
        "operationId": "install_node_ssh_pubkey_api_nodes__node_id__install_ssh_pubkey_post",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nodes/{node_id}/config": {
      "put": {
        "summary": "Update Node Config",
        "operationId": "update_node_config_api_nodes__node_id__config_put",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Node Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeConfigInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nodes/{node_id}": {
      "delete": {
        "summary": "Delete Node",
        "operationId": "delete_node_api_nodes__node_id__delete",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nodes/{node_id}/shutdown": {
      "post": {
        "summary": "Shutdown Node",
        "operationId": "shutdown_node_api_nodes__node_id__shutdown_post",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nodes/{node_id}/reboot": {
      "post": {
        "summary": "Reboot Node",
        "operationId": "reboot_node_api_nodes__node_id__reboot_post",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nodes/{node_id}/trigger-agent-update": {
      "post": {
        "summary": "Trigger Agent Update",
        "operationId": "trigger_agent_update_api_nodes__node_id__trigger_agent_update_post",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nodes/{node_id}/wake": {
      "post": {
        "summary": "Wake Node",
        "operationId": "wake_node_api_nodes__node_id__wake_post",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/node-join-tokens": {
      "get": {
        "summary": "Node Join Tokens",
        "operationId": "node_join_tokens_api_node_join_tokens_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Node Join Token",
        "operationId": "create_node_join_token_api_node_join_tokens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinTokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/node-join-tokens/{token_id}": {
      "delete": {
        "summary": "Delete Node Join Token",
        "operationId": "delete_node_join_token_api_node_join_tokens__token_id__delete",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JoinTokenCreate": {
        "properties": {
          "expected_hostname": {
            "type": "string",
            "title": "Expected Hostname",
            "default": ""
          },
          "node_group": {
            "type": "string",
            "title": "Node Group",
            "default": "unassigned"
          },
          "server_url": {
            "type": "string",
            "title": "Server Url",
            "default": ""
          },
          "expires_in_hours": {
            "type": "integer",
            "title": "Expires In Hours",
            "default": 24
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "title": "JoinTokenCreate"
      },
...
      "NodeConfigInput": {
        "properties": {
          "node_type": {
            "type": "string",
            "enum": [
              "compute",
              "storage",
              "app",
              "mixed"
            ],
            "title": "Node Type",
            "default": "compute"
          },
          "schedulable": {
            "type": "boolean",
            "title": "Schedulable",
            "default": true
          },
          "maintenance": {
            "type": "boolean",
            "title": "Maintenance",
            "default": false
          },
          "max_containers": {
            "type": "integer",
            "title": "Max Containers",
            "default": 8
          },
          "max_running_containers": {
            "type": "integer",
            "title": "Max Running Containers",
            "default": 8
          },
          "max_gpu_shared_containers": {
            "type": "integer",
            "title": "Max Gpu Shared Containers",
            "default": 4
          },
          "allow_gpu_sharing": {
            "type": "boolean",
            "title": "Allow Gpu Sharing",
            "default": true
          },
          "max_cpu_per_container": {
            "type": "integer",
            "title": "Max Cpu Per Container",
            "default": 0
          },
          "max_memory_gb_per_container": {
            "type": "integer",
            "title": "Max Memory Gb Per Container",
            "default": 0
          },
          "max_disk_gb_per_container": {
            "type": "integer",
            "title": "Max Disk Gb Per Container",
            "default": 0
          },
          "root_disk_gb": {
            "type": "integer",
            "title": "Root Disk Gb",
            "default": 0
          },
          "reserved_memory_gb": {
            "type": "integer",
            "title": "Reserved Memory Gb",
            "default": 0
          },
          "reserved_disk_gb": {
            "type": "integer",
            "title": "Reserved Disk Gb",
            "default": 0
          },
          "allow_port_mapping": {
            "type": "boolean",
            "title": "Allow Port Mapping",
            "default": true
          },
          "max_ports_per_container": {
            "type": "integer",
            "title": "Max Ports Per Container",
            "default": 8
          },
          "scheduler_weight": {
            "type": "integer",
            "title": "Scheduler Weight",
            "default": 0
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "wol_mac": {
            "type": "string",
            "title": "Wol Mac",
            "default": ""
          },
          "wol_broadcast": {
            "type": "string",
            "title": "Wol Broadcast",
            "default": "255.255.255.255"
          },
          "ssh_user": {
            "type": "string",
            "title": "Ssh User",
            "default": "root"
          },
          "ssh_port": {
            "type": "integer",
            "title": "Ssh Port",
            "default": 22
          },
          "sync_ip": {
            "type": "string",
            "title": "Sync Ip",
            "default": ""
          },
          "sync_ssh_port": {
            "type": "integer",
            "title": "Sync Ssh Port",
            "default": 0
          },
          "resource_cache_base": {
            "type": "string",
            "title": "Resource Cache Base",
            "default": ""
          },
          "shared_storage_mode": {
            "type": "string",
            "enum": [
              "inherit",
              "disabled",
              "enabled"
            ],
            "title": "Shared Storage Mode",
            "default": "inherit"
          }
        },
        "type": "object",
        "title": "NodeConfigInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}